fix: 화면 타입 간 전환 시 오류 수정 #5 #6
- listCollectionView, gridCollectionView 프로퍼티 추가
- 각 view의 dataSource 프로퍼티를 생성
- VC의 view에 각 컬렉션뷰를 주입하는 방식으로 화면전환
- 전역변수 viewType 생성을 통한 UI관련 메서드 분기 처

+++ OpenMarket/OpenMarket/Model/Product.swift
@@ -6,9 +6,9 @@ struct Product: Decodable, Hashable {
     var name: String
     var thumbnail: String
     var currency: Currency
-    var price: Int
-    var bargainPrice: Int
-    var discountedPrice: Int
+    var price: Double
+    var bargainPrice: Double
+    var discountedPrice: Double
     var stock: Int
     var createdAt: Date
     var issuedAt: Date
